Function that returns the sum of all its parameters

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -13,7 +13,7 @@ int sum_them_all(const unsigned int n, ...)
 	va_start(vars, n);
 
 	sum = 0;
-	for (i = 0; i < n; i++)
+	for (i = 0; i < (int)n; i++)
 	{
 		sum += va_arg(vars, int);
 	}
